requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.10.1
google-generativeai==0.3.2
python-dotenv==1.0.0
chromadb==0.4.15
beautifulsoup4==4.12.2
requests==2.31.0
PyPDF2==3.0.1
sentence-transformers==2.3.1
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
python-docx